Integrate More Mob Heads and Even More Mob Heads
More Mob Heads is implemented as a set of loot tables to inject to the default
loot tables. The only difference between More Mob Heads and the injected loot
tables is that the shulkers don't drop additional shells. The Ender Dragon
still drops elytras and dragon eggs.

Even More Mob Heads is implemented as a `LivingDropsEvent` listener and the
mob heads are hardcoded.